Extract controlller current_user stuff to a concern
